requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
uvicorn>=0.27.0